pyproject.toml
[project]
name = "good-boy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pyfirmata2>=2.5.1",
    "python-dotenv>=1.2.1",
]
